refactor: API updates, modularization, rustfmt, and doc improvements
